getWfdbClass: don't cache configuration settings.
Previously, getWfdbClass would cache the result of wfdbloadlib, so if
the configuration were changed, the old configuration (particularly
DEBUG_LEVEL and NETWORK_WAIT_TIME) would continue to be used.

We want to be able to change the configuration at runtime, so don't
cache the result of wfdbloadlib.  (wfdbloadlib itself will cache
things when appropriate, so calling wfdbloadlib many times should be
harmless.)

Instead, call wfdbloadlib every time and apply the current
configuration settings to the Wfdbexec object.

Note that this still won't really work yet, because most other
functions cache the result of getWfdbClass.