getWfdbClass: maintain a cache of Wfdbexec objects.
For efficiency, we want to avoid recreating the same Java objects many
times (and therefore, previously, each toolbox function would cache
the result of getWfdbClass.)  We do, however, want to apply the
current configuration settings each time.

Therefore, keep a cache of Wfdbexec objects as a persistent struct in
getWfdbClass, so that getWfdbClass('rdsamp') always returns the same
object.